requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pandas
openpyxl
python-multipart
python-dotenv
mysql-connector-python